refactor(stm): extract witness assignment into witness_assignments.rs
